+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 3.9.0 (in progress)
  - Put SHTCX to sleep after reads
  - Poll data ready before reading TMP117 in single-shot mode
- - Fix TMP117 
+ - Add delay to TMP117 after soft reset
  - Add sink-only GPIO CFG options for I2C.
  - Add RD_WARNING_DEPRECATED error code